+++ src/compiler/tokens.c
@@ -191,6 +191,10 @@ const char *token_type_to_string(TokenType type)
 			return "asm";
 		case TOKEN_ASSERT:
 			return "assert";
+		case TOKEN_ATTRGROUP:
+			return "attrgroup";
+		case TOKEN_ATTRMACRO:
+			return "attrmacro";
 		case TOKEN_ATTRDEF:
 			return "attrdef";
 		case TOKEN_BITSTRUCT:
@@ -205,6 +209,10 @@ const char *token_type_to_string(TokenType type)
 			return "const";
 		case TOKEN_CONSTDEF:
 			return "constdef";
+		case TOKEN_CENUM:
+			return "cenum";
+		case TOKEN_CONSTSET:
+			return "constset";
 		case TOKEN_CONTINUE:
 			return "continue";
 		case TOKEN_DEFAULT:
@@ -221,8 +229,14 @@ const char *token_type_to_string(TokenType type)
 			return "extern";
 		case TOKEN_FALSE:
 			return "false";
+		case TOKEN_EXCUSE:
+			return "excuse";
 		case TOKEN_FAULTDEF:
 			return "faultdef";
+		case TOKEN_FAULTCONST:
+			return "faultconst";
+		case TOKEN_FAULTSET:
+			return "faultset";
 		case TOKEN_FOR:
 			return "for";
 		case TOKEN_FOREACH:
@@ -263,6 +277,8 @@ const char *token_type_to_string(TokenType type)
 			return "true";
 		case TOKEN_TRY:
 			return "try";
+		case TOKEN_DISTINCT:
+			return "distinct";
 		case TOKEN_TYPEDEF:
 			return "typedef";
 		case TOKEN_TYPEID: